# backend/api/base/tracker_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v3.19.6
// source: api/base/tracker.proto

package base

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationTrackerBatch = "/base.Tracker/Batch"
const OperationTrackerList = "/base.Tracker/List"

type TrackerHTTPServer interface {
	// Batch Batch 批量新增埋点数据
	Batch(context.Context, *BatchRequest) (*BatchReply, error)
	// List List 分页查询埋点数据
	List(context.Context, *ListRequest) (*ListReply, error)
}

func RegisterTrackerHTTPServer(s *http.Server, srv TrackerHTTPServer) {
	r := s.Route("/")
	r.POST("/tracker/batch", _Tracker_Batch0_HTTP_Handler(srv))
	r.GET("/tracker/list", _Tracker_List0_HTTP_Handler(srv))
}

func _Tracker_Batch0_HTTP_Handler(srv TrackerHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in BatchRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTrackerBatch)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Batch(ctx, req.(*BatchRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*BatchReply)
		return ctx.Result(200, reply)
	}
}

func _Tracker_List0_HTTP_Handler(srv TrackerHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTrackerList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*ListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListReply)
		return ctx.Result(200, reply)
	}
}

type TrackerHTTPClient interface {
	// Batch Batch 批量新增埋点数据
	Batch(ctx context.Context, req *BatchRequest, opts ...http.CallOption) (rsp *BatchReply, err error)
	// List List 分页查询埋点数据
	List(ctx context.Context, req *ListRequest, opts ...http.CallOption) (rsp *ListReply, err error)
}

type TrackerHTTPClientImpl struct {
	cc *http.Client
}

func NewTrackerHTTPClient(client *http.Client) TrackerHTTPClient {
	return &TrackerHTTPClientImpl{client}
}

// Batch Batch 批量新增埋点数据
func (c *TrackerHTTPClientImpl) Batch(ctx context.Context, in *BatchRequest, opts ...http.CallOption) (*BatchReply, error) {
	var out BatchReply
	pattern := "/tracker/batch"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTrackerBatch))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List List 分页查询埋点数据
func (c *TrackerHTTPClientImpl) List(ctx context.Context, in *ListRequest, opts ...http.CallOption) (*ListReply, error) {
	var out ListReply
	pattern := "/tracker/list"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTrackerList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}