# backend/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Tracker API
    description: The tracker service definition.
    version: 0.0.1
paths:
    /tracker/batch:
        post:
            tags:
                - Tracker
            description: Batch 批量新增埋点数据
            operationId: Tracker_Batch
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BatchRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BatchReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /tracker/list:
        get:
            tags:
                - Tracker
            description: List 分页查询埋点数据
            operationId: Tracker_List
            parameters:
                - name: pageNum
                  in: query
                  description: 页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每页数量
                  schema:
                    type: integer
                    format: int32
                - name: appId
                  in: query
                  description: 应用ID
                  schema:
                    type: string
                - name: deviceId
                  in: query
                  description: 设备ID
                  schema:
                    type: string
                - name: userId
                  in: query
                  description: 用户ID
                  schema:
                    type: string
                - name: type
                  in: query
                  description: 类型
                  schema:
                    type: string
                - name: pageUrl
                  in: query
                  description: 页面URL
                  schema:
                    type: string
                - name: startTime
                  in: query
                  description: 开始时间
                  schema:
                    type: string
                - name: endTime
                  in: query
                  description: 结束时间
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        BatchReply:
            type: object
            properties:
                successCount:
                    type: integer
                    description: 成功数量
                    format: int32
                failedCount:
                    type: integer
                    description: 失败数量
                    format: int32
            description: BatchReply 批量新增响应
        BatchRequest:
            type: object
            properties:
                logs:
                    type: array
                    items:
                        $ref: '#/components/schemas/TrackerMessage'
                    description: 埋点数据列表
            description: BatchRequest 批量新增请求
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        ListReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/TrackerMessage'
                    description: 埋点数据列表
                total:
                    type: integer
                    description: 总数量
                    format: int32
            description: ListReply 分页查询响应
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        TrackerMessage:
            type: object
            properties:
                appId:
                    type: string
                    description: 应用ID
                deviceId:
                    type: string
                    description: 设备ID
                userId:
                    type: string
                    description: 用户ID
                timestamp:
                    type: string
                    description: 时间戳
                userAgent:
                    type: string
                    description: 用户代理
                pageUrl:
                    type: string
                    description: 页面URL
                type:
                    type: string
                    description: 类型
                data:
                    type: string
                    description: 数据
            description: Tracker 埋点数据消息
tags:
    - name: Tracker